agents/patchbot/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-patchbot"
version = "1.0.0"
description = "AI agent for autonomous CI/CD failure resolution"
authors = [
    {name = "Sentinel Team"}
]
requires-python = ">=3.11"
dependencies = [
    "sentinel-agent-sdk",
    "anthropic>=0.18.0",
    "gitpython>=3.1.40",
    "pygithub>=2.1.1",
    "tenacity>=8.2.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["patchbot*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
